+++ tests/callbacks/test_early_stopping.py
@@ -386,7 +386,7 @@ _NO_WIN = dict(marks=RunIf(skip_windows=True))
 
 
 @pytest.mark.parametrize(
-    "callbacks, expected_stop_epoch, check_on_train_epoch_end, accelerator, num_processes",
+    "callbacks, expected_stop_epoch, check_on_train_epoch_end, strategy, num_processes",
     [
         ([EarlyStopping("abc"), EarlyStopping("cba", patience=3)], 3, False, None, 1),
         ([EarlyStopping("cba", patience=3), EarlyStopping("abc")], 3, False, None, 1),
@@ -407,7 +407,7 @@ def test_multiple_early_stopping_callbacks(
     callbacks: List[EarlyStopping],
     expected_stop_epoch: int,
     check_on_train_epoch_end: bool,
-    accelerator: Optional[str],
+    strategy: Optional[str],
     num_processes: int,
 ):
     """Ensure when using multiple early stopping callbacks we stop if any signals we should stop."""
@@ -419,7 +419,7 @@ def test_multiple_early_stopping_callbacks(
         callbacks=callbacks,
         overfit_batches=0.20,
         max_epochs=20,
-        accelerator=accelerator,
+        strategy=strategy,
         num_processes=num_processes,
     )
     trainer.fit(model)